Add files via upload
Added functionality for CTD (Counter Current Values)
Fixed formatting error with get_sd.

SD started with an address of 361440, and Modbus uses 65536 coils. Thus, having an address with any number larger than 65536 creates error in the code. To fix this, the first digit was removed, giving the new address a number of 61440. This same concept was applied to CTD.